Rename instrument config folder
