+++ src/beamlime/config/instruments/bifrost.py
