+++ src/beamlime/config/instruments/dream.py
