+++ src/beamlime/config/instruments/dummy.py
