+++ src/beamlime/config/instruments/loki.py
