+++ src/beamlime/config/instruments/nmx.py
