+++ src/beamlime/services/dashboard.py
@@ -14,8 +14,8 @@ from dash.exceptions import PreventUpdate
 from beamlime import Service, ServiceBase
 from beamlime.config import config_names, models
 from beamlime.config.config_loader import load_config
+from beamlime.config.instruments import get_config
 from beamlime.config.models import ConfigKey
-from beamlime.config.raw_detectors import get_config
 from beamlime.config.streams import stream_kind_to_topic
 from beamlime.core.config_service import ConfigService
 from beamlime.core.message import StreamKind, compact_messages
