+++ tests/services/detector_data_test.py
@@ -8,7 +8,7 @@ import pytest
 from streaming_data_types import eventdata_ev44
 
 from beamlime import StreamKind
-from beamlime.config.raw_detectors import available_instruments, get_config
+from beamlime.config.instruments import available_instruments, get_config
 from beamlime.config.streams import stream_kind_to_topic
 from beamlime.core.handler import source_name
 from beamlime.fakes import FakeMessageSink
