+++ tests/services/timeseries_test.py
@@ -10,7 +10,7 @@ import scipp as sc
 from streaming_data_types import logdata_f144
 
 from beamlime import StreamKind
-from beamlime.config.raw_detectors import available_instruments
+from beamlime.config.instruments import available_instruments
 from beamlime.config.streams import stream_kind_to_topic
 from beamlime.fakes import FakeMessageSink
 from beamlime.kafka.message_adapter import FakeKafkaMessage, KafkaMessage
